PHOENIX-5709 Move index rebuild code to IndexRebuildRegionScanner
